First working version
This is the first semi working version of pwld, the per window
layout daemon. pwld is a simple program which basically keeps a
std::map<Window, layout> for X windows. It then listens for changes
in the properties of the root window. By sending signals to the
daemon the user can change the keyboard layout.